src/move_simplified/requirements.txt
pyserial
fastapi
uvicorn
opencv-python
numpy
Pillow
requests
